pipeline: don't throw exceptions for unknown errors
These are often recoverable. Also, fixed bug where we were looking for
error messages on the bus in the case of failed state changes, which you shouldn't
do if you have already installed a bus callback. This is why in some cases, we would
get extra "Could not add watch signal to bus" erros.